+++ src/items/Brush.java
@@ -9,9 +9,9 @@ public class Brush extends Item {
 	
 	public final static String BRUSH_NAME = "Brush";		
 	public final static String BRUSH_DESC = "For brushing";		
-	public final static double BRUSH_PRICE = 2;			
+	public final static float BRUSH_PRICE = 2;			
 	public final static String BRUSH_EFFECT = "happiness";	
-	public final static double BRUSH_MULT = 2;		
+	public final static float BRUSH_MULT = 2;		
 	
 	public Brush() {
 		super(BRUSH_NAME, BRUSH_DESC, BRUSH_PRICE, BRUSH_EFFECT, BRUSH_MULT);
