+++ src/items/GolemRose.java
@@ -9,9 +9,9 @@ public class GolemRose extends Item {
 	
 	public final static String ROSE_NAME = "Golem's Rose";		
 	public final static String ROSE_DESC = "It comes from the northern iron farms";		
-	public final static double ROSE_PRICE = 2;				
+	public final static float ROSE_PRICE = 2;				
 	public final static String ROSE_EFFECT = "Income";		
-	public final static double ROSE_MULT = 2;	
+	public final static float ROSE_MULT = 2;	
 	
 	public GolemRose() {
 		super(ROSE_NAME, ROSE_DESC, ROSE_PRICE, ROSE_EFFECT, ROSE_MULT);
